+++ routes/players.js
@@ -51,7 +51,7 @@ router.put('/:id', function (req, res, next) {
 
 /* Get specific statistics for a given player */
 router.get('/:id/statistics', function (req, res, next) {
-    var playerId = req.params.id;
+    const playerId = req.params.id;
     axios.all([
         axios.get(`${req.app.locals.kcapp.api}/player`),
         axios.get(`${req.app.locals.kcapp.api}/player/${playerId}`),
@@ -59,8 +59,11 @@ router.get('/:id/statistics', function (req, res, next) {
         axios.get(`${req.app.locals.kcapp.api}/player/${playerId}/statistics/previous`),
         axios.get(`${req.app.locals.kcapp.api}/player/${playerId}/progression`),
         axios.get(`${req.app.locals.kcapp.api}/player/${playerId}/checkouts`),
-        axios.get(`${req.app.locals.kcapp.api}/player/${playerId}/tournament`)
-    ]).then(axios.spread((players, player, statistics, previous, progression, checkouts, tournament) => {
+        axios.get(`${req.app.locals.kcapp.api}/player/${playerId}/tournament`),
+        axios.get(`${req.app.locals.kcapp.api}/player/${playerId}/badges`),
+        axios.get(`${req.app.locals.kcapp.api}/badge`),
+        axios.get(`${req.app.locals.kcapp.api}/tournament`)
+    ]).then(axios.spread((players, player, statistics, previous, progression, checkouts, playerTournament, playerBadges, badges, tournaments) => {
         res.marko(playerTemplate, {
             players: players.data,
             player: player.data,
@@ -68,7 +71,10 @@ router.get('/:id/statistics', function (req, res, next) {
             previous_statistics: previous.data,
             progression: progression.data,
             checkouts: checkouts.data,
-            tournament_standings: tournament.data,
+            tournament_standings: playerTournament.data,
+            player_badges: playerBadges.data,
+            badges: badges.data,
+            tournaments: tournaments.data,
             locals: req.app.locals
         });
     })).catch(error => {