+++ src/pages/player/components/player/components/tab-badges/components/badge/badge.component.js
@@ -1,0 +1,7 @@
+module.exports = {
+    onCreate(input) {
+        this.state = { 
+            id: input.badge.filename.substring(0, input.badge.filename.lastIndexOf('.'))
+        }
+    }
+}