+++ src/pages/tournament-admin/components/tournament-admin/tournament-admin.component.js
@@ -1,14 +1,22 @@
-var moment = require("moment");
+const moment = require("moment");
+const _ = require("underscore");
 
 module.exports = {
     onCreate(input) {
-        for (var i = 0; i < input.metadata.length; i++) {
-            var metadata = input.metadata[i];
-            var match = input.matches[metadata.match_id];
+        for (let i = 0; i < input.metadata.length; i++) {
+            let metadata = input.metadata[i];
+            let match = input.matches[metadata.match_id];
             match.start_time = moment(match.created_at).format('YYYY-MM-DD HH:mm');
         }
+        const matches = _.reject(Object.values(input.matches), (match) => match.is_players_decided);
         this.state = {
-            matches: input.matches
+            matchesMap: input.matches,
+            matches: matches
         }
+    },
+    
+    onShowModal(matchId) {
+        const match = this.state.matchesMap[matchId];
+        this.getComponent('set-score-modal').setMatch(match, false);
     }
 }